requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.78.0
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
dateparser==1.3.0
distro==1.9.0
docstring_parser==0.17.0
Flask==3.1.2
Flask-SocketIO==5.6.0
frozenlist==1.8.0
h11==0.16.0
huggingface_hub==0.34.3
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
openai==2.17.0
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
tavily==1.1.0
tiktoken==0.12.0
tld==0.13.1
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
yarl==1.22.0
